New pool, weapon stats, and some folder organization
The pool of objects require a gameobject either inheriting from pool or having the pool component.

also the Pool manager is in the resources folder, it needs the reference for the pool object and a name so the pool knows where to storage itself.

The pool manager has a method named Item_Call that calls the type of item the name points at, the Item_Dismiss method  is to get rid of the gameobject in question to return it to the pool